requirements.txt
# ===== NÚCLEO DEL ASISTENTE =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==1.5.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.1

# ===== PROCESAMIENTO DE DATOS =====
gspread==5.8.0
google-auth==2.17.3
openpyxl==3.1.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# ===== MACHINE LEARNING & AI =====
scikit-learn==1.2.2
prophet==1.1.1
statsmodels==0.13.5
tensorflow==2.13.0
openai==1.3.0
google-generativeai==0.3.0
transformers==4.35.2
torch==2.1.0

# ===== REPORTES Y VISUALIZACIÓN =====
weasyprint==55.0
jinja2==3.1.2
plotly==5.9.0
matplotlib==3.7.1
seaborn==0.13.0

# ===== ASISTENTE VIRTUAL (Voz/Audio) =====
pygame==2.5.0
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.11
Pillow==10.0.0
elevenlabs==0.2.26
whisper-openai==1.0.0
librosa==0.10.1
soundfile==0.12.1

# ===== DESARROLLO Y DEPLOYMENT =====
docker==6.1.3
paramiko==3.3.1
google-cloud-run==0.7.0
boto3==1.28.62
celery==5.3.6
redis==5.0.1

# ===== CALIDAD DE CÓDIGO =====
radon==5.1.0
pylint==2.17.5
black==23.11.0
flake8==6.1.0
isort==5.12.0

# ===== TESTING =====
pytest==7.4.0
pytest-cov==4.1.0
coverage==7.3.1
pytest-asyncio==0.21.1

# ===== SEGURIDAD Y AUTENTICACIÓN =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.2.1

# ===== UTILIDADES =====
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2023.3.post1
tqdm==4.66.1
loguru==0.7.2